+++ libft/ft_tabnew_two.c
@@ -6,13 +6,13 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/21 14:22:42 by cfatrane          #+#    #+#             */
-/*   Updated: 2017/01/27 20:29:24 by cfatrane         ###   ########.fr       */
+/*   Updated: 2017/01/28 11:53:50 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		**ft_tabnew_two(int x, int y)
+int		**ft_tabnew_two(size_t x, size_t y)
 {
 	size_t	i;
 	int		**tab;
